requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pydantic==2.6.3
slowapi==0.1.9

